Cam Updates
Fist Person: No has the ability to collide with "infinite pillars" (to
be used for trees)
Static camera: Got rid of the sit in one place model, will now orbit the
origin in a circular patern. This can be sped up, zoomed ect.. Will be
better tied to our scene once the scene is created.
Both cams have a GetLookAt()   for those who need them